ml_training/ml_requirements.txt
tensorflow>=2.10.0
numpy
pillow
pandas
matplotlib
scikit-learn
